Move database to a separate package
