+++ .env.example
@@ -1,1 +1,3 @@
+APP_HOST=0.0.0.0
+POST=8080
 DATABASE_URL=postgres://postgres:postgres@0.0.0.0:5432/google_scraper_development?sslmode=disable
