+++ cmd/api/main.go
@@ -4,7 +4,9 @@ import (
 	"fmt"
 
 	"go-google-scraper-challenge/bootstrap"
+	"go-google-scraper-challenge/database"
 	"go-google-scraper-challenge/helpers/log"
+	"go-google-scraper-challenge/lib/services/oauth"
 
 	"github.com/gin-gonic/gin"
 	"github.com/spf13/viper"
@@ -13,9 +15,11 @@ import (
 func main() {
 	bootstrap.LoadConfig()
 
-	bootstrap.InitDatabase(bootstrap.GetDatabaseURL())
+	database.InitDatabase(database.GetDatabaseURL())
 
 	r := bootstrap.SetupRouter()
+
+	oauth.SetUpOauth()
 
 	err := r.Run(getAppPort())
 	if err != nil {
